requirements.txt
openpyxl>=3.1.5
Flask>=2.0
python-dotenv
requests>=2.32.5
Flask-JWT-Extended>=4.4.4
flask_apscheduler>=1.13.1
pymongo>=4.0
Flask[async]
matplotlib
flask_cors
pydantic
email-validator
